Remove SpanIdImpl. SpanId becomes implementation.
SpanId becomes implementation instead of interface.
