+++ brave-interfaces/src/test/java/com/github/kristofa/brave/SpanIdTest.java
@@ -1,31 +1,29 @@
 package com.github.kristofa.brave;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
 
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Optional;
 
-public class SpanIdImplTest {
+import static org.junit.Assert.*;
+
+public class SpanIdTest {
 
     private final static long TRACEID = 10;
     private final static long SPANID = 11;
     private final static Long PARENT_SPANID = new Long(12);
 
-    private SpanIdImpl spanIdImpl;
+    private SpanId spanIdImpl;
 
     @Before
     public void setup() {
-        spanIdImpl = new SpanIdImpl(TRACEID, SPANID, Optional.of(PARENT_SPANID));
+        spanIdImpl = new SpanId(TRACEID, SPANID, Optional.of(PARENT_SPANID));
     }
 
     @Test
     public void testSpanIdNullParentId() {
-        final SpanIdImpl spanIdImpl2 = new SpanIdImpl(TRACEID, SPANID, Optional.empty());
+        final SpanId spanIdImpl2 = new SpanId(TRACEID, SPANID, Optional.empty());
         assertNull(spanIdImpl2.getParentSpanId());
     }
 
@@ -52,7 +50,7 @@ public class SpanIdImplTest {
 
     @Test
     public void testHashCode() {
-        final SpanIdImpl equalSpanId = new SpanIdImpl(TRACEID, SPANID, Optional.of(PARENT_SPANID));
+        final SpanId equalSpanId = new SpanId(TRACEID, SPANID, Optional.of(PARENT_SPANID));
         assertEquals(spanIdImpl.hashCode(), equalSpanId.hashCode());
     }
 
@@ -62,28 +60,29 @@ public class SpanIdImplTest {
         assertFalse(spanIdImpl.equals(null));
         assertFalse(spanIdImpl.equals(new String()));
 
-        final SpanIdImpl equalSpanId = new SpanIdImpl(TRACEID, SPANID, Optional.of(PARENT_SPANID));
+        final SpanId equalSpanId = new SpanId(TRACEID, SPANID, Optional.of(PARENT_SPANID));
         assertTrue(spanIdImpl.equals(equalSpanId));
 
-        final SpanIdImpl nonEqualSpanId = new SpanIdImpl(TRACEID + 1, SPANID, Optional.of(PARENT_SPANID));
-        final SpanIdImpl nonEqualSpanId2 = new SpanIdImpl(TRACEID, SPANID + 1, Optional.of(PARENT_SPANID));
-        final SpanIdImpl nonEqualSpanId3 = new SpanIdImpl(TRACEID, SPANID, Optional.of(PARENT_SPANID + 1));
+        final SpanId nonEqualSpanId = new SpanId(TRACEID + 1, SPANID, Optional.of(PARENT_SPANID));
+        final SpanId nonEqualSpanId2 = new SpanId(TRACEID, SPANID + 1, Optional.of(PARENT_SPANID));
+        final SpanId nonEqualSpanId3 = new SpanId(TRACEID, SPANID, Optional.of(PARENT_SPANID + 1));
 
         assertFalse(spanIdImpl.equals(nonEqualSpanId));
         assertFalse(spanIdImpl.equals(nonEqualSpanId2));
         assertFalse(spanIdImpl.equals(nonEqualSpanId3));
     }
 
+
+
     @Test
     public void testToString() {
         assertEquals("[trace id: " + TRACEID + ", span id: " + SPANID + ", parent span id: Optional[" + PARENT_SPANID + "]]",
-            spanIdImpl.toString());
+                spanIdImpl.toString());
     }
 
     @Test
     public void testToStringNullParent() {
-        final SpanIdImpl spanIdImpl2 = new SpanIdImpl(TRACEID, SPANID, Optional.empty());
+        final SpanId spanIdImpl2 = new SpanId(TRACEID, SPANID, Optional.empty());
         assertEquals("[trace id: " + TRACEID + ", span id: " + SPANID + ", parent span id: Optional.empty]", spanIdImpl2.toString());
     }
-
 }
